:recycle: (Venue) Split venue to add content in a separate file
